+++ src/anemoi/transform/filters/rescale.py
@@ -6,80 +6,64 @@
 # In applying this licence, ECMWF does not waive the privileges and immunities
 # granted to it by virtue of its status as an intergovernmental organisation
 # nor does it submit to any jurisdiction.
-
-
-from collections.abc import Iterator
+from abc import ABC
+from abc import abstractmethod
+from typing import Callable
 
 import earthkit.data as ekd
 
+from anemoi.transform.filter import SingleFieldFilter
 from anemoi.transform.filters import filter_registry
-from anemoi.transform.filters.matching import MatchingFieldsFilter
-from anemoi.transform.filters.matching import matching
 
 
-class Rescale(MatchingFieldsFilter):
-    """A filter to rescale a parameter from a scale and an offset, and back."""
-
-    @matching(
-        select="param",
-        forward=("param",),
-        backward=("param",),
-    )
-    def __init__(
-        self,
-        *,
-        scale: float,
-        offset: float,
-        param: str,
-    ) -> None:
-        """Parameters
-        -------------
-        scale : float
-            The scale factor.
-        offset : float
-            The offset value.
-        param : str
-            The parameter to be rescaled.
-        """
-
+class Rescaler:
+    def __init__(self, scale, offset):
         self.scale = scale
         self.offset = offset
-        self.param = param
 
-    def forward_transform(self, param: ekd.Field) -> Iterator[ekd.Field]:
-        """Apply the forward transformation (x to ax+b).
+    def forward(self, x):
+        return x * self.scale + self.offset
 
-        Parameters
-        ----------
-        param : ekd.Field
-            The input data to be transformed.
-
-        Returns
-        -------
-        Iterator[ekd.Field]
-            A generator yielding the transformed data.
-        """
-        rescaled = param.to_numpy() * self.scale + self.offset
-        yield self.new_field_from_numpy(rescaled, template=param, param=self.param)
-
-    def backward_transform(self, param: ekd.Field) -> Iterator[ekd.Field]:
-        """Apply the backward transformation (ax+b to x).
-
-        Parameters
-        ----------
-        param : ekd.Field
-            The input data to be transformed.
-
-        Returns
-        -------
-        Iterator[ekd.Field]
-            A generator yielding the transformed data.
-        """
-        descaled = (param.to_numpy() - self.offset) / self.scale
-        yield self.new_field_from_numpy(descaled, template=param, param=self.param)
+    def backward(self, x):
+        return (x - self.offset) / self.scale
 
 
-class Convert(Rescale):
+class RescaleMixin(ABC):
+    # inheriting classes should define required_inputs (which must include param)
+    # and must define self.rescaler in prepare_filter
+    param: str
+    rescaler: Rescaler
+    # intended to be inherited from SingleFieldFilter
+    new_field_from_numpy: Callable
+
+    @abstractmethod
+    def prepare_filter(self):
+        raise NotImplementedError("prepare_filter must be implemented by subclasses.")
+
+    def forward_select(self):
+        return {"param": self.param}
+
+    def forward_transform(self, param: ekd.Field) -> ekd.Field:
+        """Apply the forward transformation (x to ax+b)."""
+        rescaled = self.rescaler.forward(param.to_numpy())
+        return self.new_field_from_numpy(rescaled, template=param, param=self.param)
+
+    def backward_transform(self, param: ekd.Field) -> ekd.Field:
+        """Apply the backward transformation (ax+b to x)."""
+        descaled = self.rescaler.backward(param.to_numpy())
+        return self.new_field_from_numpy(descaled, template=param, param=self.param)
+
+
+class Rescale(RescaleMixin, SingleFieldFilter):
+    """A filter to rescale a parameter from a scale and an offset, and back."""
+
+    required_inputs = ("scale", "offset", "param")
+
+    def prepare_filter(self):
+        self.rescaler = Rescaler(self.scale, self.offset)
+
+
+class Convert(RescaleMixin, SingleFieldFilter):
     """A filter to convert a parameter in a given unit to another unit, and back.
 
     This filter uses :mod:`cfunits` (see the `cfunits documentation <https://ncas-cms.github.io/cfunits/>`_)
@@ -102,26 +86,19 @@ class Convert(Rescale):
 
     """
 
-    def __init__(self, *, unit_in: str, unit_out: str, param: str) -> None:
-        """Parameters
-        -------------
-        unit_in : str
-            The input unit.
-        unit_out : str
-            The output unit.
-        param : str
-            The parameter to be converted.
-        """
+    required_inputs = ("unit_in", "unit_out", "param")
+
+    def prepare_filter(self):
         from cfunits import Units
 
-        u0 = Units(unit_in)
-        u1 = Units(unit_out)
+        u0 = Units(self.unit_in)
+        u1 = Units(self.unit_out)
         x1, x2 = 0.0, 1.0
         y1, y2 = Units.conform([x1, x2], u0, u1)
-        a = (y2 - y1) / (x2 - x1)
-        b = y1 - a * x1
+        scale = (y2 - y1) / (x2 - x1)
+        offset = y1 - scale * x1
 
-        super().__init__(scale=a, offset=b, param=param)
+        self.rescaler = Rescaler(scale, offset)
 
 
 filter_registry.register("rescale", Rescale)